+++ CHANGELOG.md
@@ -230,6 +230,8 @@
 
 * [x] Wortox health changed to 120 from 200
 
+* [x] Added options to reduce the Bee Queen and Toadstool's health, due to their raised importance.
+
 ### To check current things worked on: Visit [#mod_roadmap](https://discordapp.com/channels/623649948130344960/624590758959382539) on Discord
 
 ### To vote on features: Visit [#mod_polls](https://discordapp.com/channels/623649948130344960/623661746216763393) on Discord